housekeeping: prepare for sorted series by adjusting contexts
With the patches applied out of order, we can sometimes end up
with code applied to a different location than in upstream.  This commit
only changes context to apply the patch in the correct location.

Since the next commit will touch many, many files, this commit contains
the changes that have visible impact on the end result and is a much
smaller set.

There are no code changes, just movement.