arm64: hardcode rodata_enabled=true earlier in the series (bsc#1114763).
In the initial backports that depend on rodata_enabled, we shift to to
be a literal true, which causes compilation issues.  Later, we add the
placeholder rodata_enabled variable to fix that.  This commit moves it prior
to applying any of the patches that use it, avoiding the compilation issue
entirely.  Also mark it const.